wallet_managment/requirements.txt
fastapi
pytest
databases
sqlalchemy
python-dotenv
pydantic
uvicorn
asyncpg
httpx
pytest-asyncio
